+++ scm_helper/conduct.py
@@ -1,5 +1,6 @@
 """SCM Conduct."""
 from scm_helper.config import (
+    A_DATEAGREED,
     A_GUID,
     A_MEMBERS,
     C_CONDUCT,
@@ -96,12 +97,12 @@ class Conduct(Entity):
         # This approach breaks the model. Oh well.
         for member in self.data[A_MEMBERS]:
 
-            if member["DateAgreed"]:
+            if member[A_DATEAGREED]:
                 continue
 
             person = self.scm.members.by_guid[member[A_GUID]]
             if person.confirmed_date:  # Will get a not confirmed error later in not set
-                issue(person, E_NO_CONDUCT_DATE, self.name)
+                issue(person, E_NO_CONDUCT_DATE, self.name, 0, person.first_group)
                 codes = get_config(self.scm, C_LISTS, C_CONDUCT)
                 if codes:
                     for code in codes: