+++ scm_helper/issue.py
@@ -614,10 +614,11 @@ def print_dict(xdict, reports):
             inner_match = False
             first = True
             length = len(xdict[key1][key2])
+            # pylint: disable=bad-continuation
+            # black insists
             for xissue in sorted(
-                    xdict[key1][key2],
-                    key=lambda x: "None" if x[0] is None else x[0]
-                ):
+                xdict[key1][key2], key=lambda x: "None" if x[0] is None else x[0]
+            ):
                 val1, val2, rpt, rev, _ = xissue
                 if (first and rev) or (first and (length > 1)):
                     to_print += "\n"